Move parsePclntabFuncsRaw to test-only file
parsePclntabFuncsRaw is used only to supply uncorrected pclntab entries
to offset-detection tests; it has no production callers.  Move it to
pclntab_parser_helpers_test.go (//go:build test || integration) so it is
compiled only during testing and does not appear in the production source.

Also remove the now-redundant parsePclntabRaw wrapper from the integration
test file; callers use parsePclntabFuncsRaw directly.
